+++ gslab_scons/misc.py
@@ -9,7 +9,6 @@ import datetime
 import _exception_classes
 from size_warning import issue_size_warnings
 
-
 def scons_debrief(target, source, env):
     '''Execute functions after SCons has built all targets'''
     # Log the state of the repo
@@ -235,6 +234,51 @@ def lyx_scan(node, env, path):
                  for source in src_find.findall(contents)]
 
     return SOURCE
+
+
+def load_yaml_value(path, key):
+    '''
+    Load the yaml value indexed by the key argument in the file
+    specified by the path argument.
+    '''
+    import yaml
+
+    if key == "stata_executable":
+        prompt = "Enter %s value or None to search for defaults: "
+    else:
+        prompt = "Enter %s value: "
+
+    # Check if file exists and is not corrupted. If so, load yaml contents.
+    yaml_contents = None
+    if os.path.isfile(path):
+        try:
+            yaml_contents = yaml.load(open(path, 'rU'))
+            if not isinstance(yaml_contents, dict):
+                raise yaml.scanner.ScannerError()
+
+        except yaml.scanner.ScannerError:
+            message  = "%s is a corrupted yaml file. Delete file and recreate? (y/n) "
+            response = str(raw_input(message % path))
+            if response.lower() == 'y':
+                os.remove(path)
+                yaml_contents = None
+            else:
+                message = "%s is a corrupted yaml file. Please fix." % path
+                raise _exception_classes.PrerequisiteError(message)
+
+    # If key exists, return value. Otherwise, add key-value to file.
+    try:
+        if yaml_contents[key] == "None":
+            return None
+        else:
+            return yaml_contents[key]
+    except:
+        with open(path, 'ab') as f:        
+            val = str(raw_input(prompt % key))
+            if re.sub('"', '', re.sub('\'', '', val.lower())) == "none":
+                val = None
+            f.write('%s: %s\n' % (key, val))
+        return val
 
 
 def get_directory(path):
