+++ gslab_scons/tests/user-config.yaml
@@ -1,0 +1,1 @@
+stata_executable:
